requirements.txt
Pillow
DrissionPage
pydub
SpeechRecognition
python-dotenv
openai